.js to .jsx and .css to .module.scss
